use double

+++ lab02/l021.cpp
@@ -11,30 +11,30 @@
 
 class Point {
     public:
-        float x,y;
+        double x,y;
     Point(){
         x = 0;
         y = 0;
     }
-    Point(float px, float py){
+    Point(double px, double py){
         x = px; 
         y = py;
     }
 };
 
-float area(Point p1, Point p2, Point p3) 
+double area(Point p1, Point p2, Point p3) 
 { 
    return std::abs((p1.x*(p2.y-p3.y) + p2.x*(p3.y-p1.y)+ p3.x*(p1.y-p2.y))/2.0); 
 } 
 
 
 bool in_triangle(Point verts[4]){
-    float A = area(verts[0], verts[1], verts[2]);
-    float s = area(verts[0], verts[1], verts[3]) + area(verts[0], verts[3], verts[2]) + area(verts[3], verts[1], verts[2]);
+    double A = area(verts[0], verts[1], verts[2]);
+    double s = area(verts[0], verts[1], verts[3]) + area(verts[0], verts[3], verts[2]) + area(verts[3], verts[1], verts[2]);
     return A-e <= s && A+e >= s;
 }
 
-float area(int x1, int y1, int x2, int y2, int x3, int y3) 
+double area(int x1, int y1, int x2, int y2, int x3, int y3) 
 { 
    return abs((x1*(y2-y3) + x2*(y3-y1)+ x3*(y1-y2))/2.0); 
 } 
@@ -44,11 +44,11 @@ int main(int argc, char** argv){
     srand(time(0));
     Point vertices[4];
     for(int i=0;i<3;i++){
-        vertices[i] = Point(rand()/(float)RAND_MAX, rand()/(float)RAND_MAX);
+        vertices[i] = Point(rand()/(double)RAND_MAX, rand()/(double)RAND_MAX);
     }
     
     do{
-        vertices[3] = Point(rand()/(float)RAND_MAX, rand()/(float)RAND_MAX);
+        vertices[3] = Point(rand()/(double)RAND_MAX, rand()/(double)RAND_MAX);
     }while(in_triangle(vertices) == true);
     
     fout = fopen("points.txt", "w");
